Added own_grocery_list function and scrollable container
This function deletes all items without saving them. Optionally we can also program a settings button and by clicking on it the user can declare if he wants to save the deleted items or not.  I also added that you can scroll now within the container as we agreed on